requirements.txt
# LLM-Optimized Requirements for SISIMPUR
# This file contains minimal dependencies for LLM-based OCR

# Core Django and web framework
Django==5.0.9
django_browser_reload
python-dotenv

# Google APIs and AI
google_api_python_client
google_auth_oauthlib
google-generativeai

# Document processing
PyMuPDF
pdf2image
PyYAML

# Image processing (minimal - no OCR dependencies)
opencv-python-headless
Pillow

# Build and utility tools
setuptools
wheel